+++ tramex-tools/src/interface/interface_file/mod.rs
@@ -1,3 +1,4 @@
-//! Websocket module
+//! File interface module
 
 pub mod file_handler;
+pub mod utils_file;
