+++ tramex-tools/src/interface/mod.rs
@@ -1,5 +1,6 @@
 //! Interface module
 
+pub mod functions;
 pub mod interface_types;
 pub mod layer;
 pub mod log_get;
