+++ tramex-tools/tests/tests_deserialize.rs
@@ -50,7 +50,7 @@ mod tests {
         }
     }
 
-    use tramex_tools::functions::extract_hexe;
+    use tramex_tools::interface::functions::extract_hexe;
     #[test]
     fn test_extract_hexe() {
         let data = vec![
